backend/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-environ==0.4.5
django-filter==2.2.0
django-localflavor==3.0.1
djangorestframework==3.11.0
mysqlclient==1.4.6
python-stdnum==1.13
pytz==2019.3
sqlparse==0.3.1
